requirements.txt
requests==2.31.0
pandas>=2.2.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.3.0
praw>=7.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
plotly>=5.17.0
python-dotenv>=1.0.0 